+++ files/ionprinter/simulation/nyion/src/visualize.cpp
@@ -12,19 +12,19 @@
 #include <GL/glx.h>
 
 
-float opengl_zoom = 30;
-float opengl_z_extent = 10;
-float opengl_current_angle_x = 45;
-float opengl_current_angle_y = 25.0;
-float opengl_current_x_translate = 0.0;
-float opengl_current_y_translate = 0.0;
-float opengl_current_z_translate = 0;
-float opengl_delta_zoom = 0;
-float opengl_delta_angle_x = 0;
-float opengl_delta_angle_y = 0;
-float opengl_delta_x_translate = 0.0;
-float opengl_delta_y_translate = 0.0;
-float opengl_delta_z_translate = 0.0;
+double opengl_zoom = 30;
+double opengl_z_extent = 10;
+double opengl_current_angle_x = 45;
+double opengl_current_angle_y = 25.0;
+double opengl_current_x_translate = 0.0;
+double opengl_current_y_translate = 0.0;
+double opengl_current_z_translate = 0;
+double opengl_delta_zoom = 0;
+double opengl_delta_angle_x = 0;
+double opengl_delta_angle_y = 0;
+double opengl_delta_x_translate = 0.0;
+double opengl_delta_y_translate = 0.0;
+double opengl_delta_z_translate = 0.0;
 
 GLXDrawable mesh_window;
 GLXDrawable graph_window;
@@ -245,8 +245,8 @@ void draw_mesh(std::vector<std::vector<T>>& input_mesh, root_mesh_geometry mesh_
   /* -----------------------------------------------------------------------------
   Determine min and max for color scaling
   ----------------------------------------------------------------------------- */
-  float min = mesh_min(input_mesh,mesh_geometry);
-  float max = mesh_max(input_mesh,mesh_geometry);
+  double min = mesh_min(input_mesh,mesh_geometry);
+  double max = mesh_max(input_mesh,mesh_geometry);
   /* -----------------------------------------------------------------------------
   Draw cubes
   ----------------------------------------------------------------------------- */
@@ -273,7 +273,7 @@ void draw_mesh(std::vector<std::vector<T>>& input_mesh, root_mesh_geometry mesh_
           int sub_len = submesh_side_length(input_mesh[root_idx]);
 
           cube_size = (mesh_geometry.root_scale/sub_len)/OPENGL_WORLD_SCALE;
-          float fine_scale = mesh_geometry.root_scale/sub_len;
+          double fine_scale = mesh_geometry.root_scale/sub_len;
           for(int x = 0; x < sub_len; x++){
             for(int y = 0; y < sub_len; y++){
               for(int z = 0; z < sub_len; z++){
@@ -315,7 +315,7 @@ void draw_mesh(std::vector<std::vector<T>>& input_mesh, root_mesh_geometry mesh_
     }
   }
 }
-template void draw_mesh(std::vector<std::vector<float>>& input_mesh,root_mesh_geometry mesh_geometry);
+template void draw_mesh(std::vector<std::vector<double>>& input_mesh,root_mesh_geometry mesh_geometry);
 template void draw_mesh(std::vector<std::vector<int>>& input_mesh,root_mesh_geometry mesh_geometry);
 
 
@@ -335,22 +335,22 @@ void opengl_apply_camera_rotation(){
 }
 
 
-void opengl_graph_1d_vector(std::vector<float> &input, const std::string& title, int index){
-  float min = *std::min_element(input.begin(),input.end());
-  float max = *std::max_element(input.begin(),input.end());
+void opengl_graph_1d_vector(std::vector<double> &input, const std::string& title, int index){
+  double min = *std::min_element(input.begin(),input.end());
+  double max = *std::max_element(input.begin(),input.end());
 
   opengl_switch_to_graph_window();
 
-  float graph_y_position = (OPENGL_GRAPH_WINDOW_Y/ ((float)OPENGL_GRAPH_COUNT))*index + OPENGL_GRAPH_Y_OFFSET;
-  float graph_height = OPENGL_GRAPH_WINDOW_Y/((float)OPENGL_GRAPH_COUNT);
+  double graph_y_position = (OPENGL_GRAPH_WINDOW_Y/ ((double)OPENGL_GRAPH_COUNT))*index + OPENGL_GRAPH_Y_OFFSET;
+  double graph_height = OPENGL_GRAPH_WINDOW_Y/((double)OPENGL_GRAPH_COUNT);
   glRasterPos2i(OPENGL_GRAPH_X_OFFSET,graph_y_position);
   glColor4f(1.0f, 1.0f, 1.0f, 1.0f);
   glutBitmapString(GLUT_BITMAP_HELVETICA_18, (unsigned char *) title.c_str());
 
   glColor3f(1.0f, 1.0f, 0.0f);
   glBegin(GL_LINE_STRIP);
-  float graph_x_scale = (((float)OPENGL_GRAPH_WINDOW_X-OPENGL_GRAPH_X_OFFSET)/input.size());
-  float graph_y_scale = graph_height/(max-min);
+  double graph_x_scale = (((double)OPENGL_GRAPH_WINDOW_X-OPENGL_GRAPH_X_OFFSET)/input.size());
+  double graph_y_scale = graph_height/(max-min);
   for(uint32_t i = 0; i < input.size(); i++){
     glVertex2f(OPENGL_GRAPH_X_OFFSET+(graph_x_scale*i),(graph_y_position+graph_height)-(graph_y_scale*(input[i]-min)));
   }
@@ -372,8 +372,8 @@ void update_screen(){
 
 //
 // void mouse_handler(int x, int y){
-//   opengl_angle_x = ((x-mouse_previous_x)/(float)OPENGL_3D_WINDOW_X)*360.0;
-//   opengl_angle_y = ((y-mouse_previous_y)/(float)OPENGL_3D_WINDOW_Y)*360.0;
+//   opengl_angle_x = ((x-mouse_previous_x)/(double)OPENGL_3D_WINDOW_X)*360.0;
+//   opengl_angle_y = ((y-mouse_previous_y)/(double)OPENGL_3D_WINDOW_Y)*360.0;
 //   mouse_previous_x = x;
 //   mouse_previous_y = y;
 //
@@ -382,7 +382,7 @@ void update_screen(){
 
 
 
-// void TakeScreenshot(float time)
+// void TakeScreenshot(double time)
 // {
 // 	unsigned char *buffer;
 // 	char filename[200];
@@ -434,7 +434,7 @@ void update_screen(){
 
 
 
-void to_csv(std::vector<std::vector<float>> &original, root_mesh_geometry mesh_geometry){
+void to_csv(std::vector<std::vector<double>> &original, root_mesh_geometry mesh_geometry){
   std::ofstream output_file;
   output_file.open ("out.csv");
 
@@ -449,9 +449,9 @@ void to_csv(std::vector<std::vector<float>> &original, root_mesh_geometry mesh_g
             for(int y = 0; y < fine_sub_len; y++){
               for(int z = 0; z < fine_sub_len; z++){
                 int sub_idx = (fine_sub_len*fine_sub_len*z) + (fine_sub_len*y) + x;
-                float world_x = (r_x*mesh_geometry.root_scale)+(x*(mesh_geometry.root_scale/fine_sub_len));
-                float world_y = (r_y*mesh_geometry.root_scale)+(y*(mesh_geometry.root_scale/fine_sub_len));
-                float world_z = (r_z*mesh_geometry.root_scale)+(z*(mesh_geometry.root_scale/fine_sub_len));
+                double world_x = (r_x*mesh_geometry.root_scale)+(x*(mesh_geometry.root_scale/fine_sub_len));
+                double world_y = (r_y*mesh_geometry.root_scale)+(y*(mesh_geometry.root_scale/fine_sub_len));
+                double world_z = (r_z*mesh_geometry.root_scale)+(z*(mesh_geometry.root_scale/fine_sub_len));
                 // if(original[root_idx][sub_idx]){
                   output_file << world_x << ","
                               << world_y << ","
